Changes qtimer_fastrand() to return an unsigned long.
git-svn-id: https://qthreads.googlecode.com/svn/trunk@2452 9dbd018f-d1da-d342-4c11-58c73dfe904b

+++ src/qtimer/altix.c
@@ -81,9 +81,9 @@ void qtimer_start(qtimer_t q)
     q->start = vol_read_ul(timer_address);
 }
 
-long qtimer_fastrand(void)
+unsigned long qtimer_fastrand(void)
 {
-    return (long)(vol_read_ul(timer_address));
+    return vol_read_ul(timer_address);
 }
 
 void qtimer_stop(qtimer_t q)
